+++ CHANGES.rst
@@ -1,5 +1,11 @@
 .. currentmodule:: werkzeug
 
+Version 3.1.4
+-------------
+
+Unreleased
+
+
 Version 3.1.3
 -------------